fix Android compile error
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@59991 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/gui/dialog/options/model/ObjectSettingsModel.java
@@ -471,10 +471,12 @@ abstract public class ObjectSettingsModel {
     }
 
     /**
-     * @return the current alpha value of the geoElement
-     */
-	public double getAlpha() {
-		return geoElement != null ? geoElement.getAlphaValue() : 1;
+	 * @return the current alpha value of the geoElement
+	 * 
+	 *         (needs to be float for Android)
+	 */
+	public float getAlpha() {
+		return (float) (geoElement != null ? geoElement.getAlphaValue() : 1);
     }
 
     /**
